+++ spring-integration-file/src/main/java/org/springframework/integration/file/outbound/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * Inbound channel adapters for file system.
+ */
+@org.jspecify.annotations.NullMarked
+package org.springframework.integration.file.outbound;
