+++ spring-integration-file/src/test/java/org/springframework/integration/file/config/FileOutboundChannelAdapterParserTests-context.xml
@@ -64,7 +64,7 @@
 
 	<bean id="predicate" class="org.mockito.Mockito" factory-method="mock">
 		<constructor-arg
-				value="org.springframework.integration.file.FileWritingMessageHandler.MessageFlushPredicate"
+				value="org.springframework.integration.file.outbound.FileWritingMessageHandler.MessageFlushPredicate"
 				type="java.lang.Class"/>
 	</bean>
