+++ spring-integration-file/src/test/java/org/springframework/integration/file/inbound/FileReadingMessageSourceIntegrationTests-context.xml
@@ -6,8 +6,8 @@
 			https://www.springframework.org/schema/beans/spring-beans.xsd">
 
     <!-- under test -->
-    <bean id="pollableFileSource" class="org.springframework.integration.file.FileReadingMessageSource"
-          p:directory="#{T(org.springframework.integration.file.FileReadingMessageSourceIntegrationTests).inputDir}"
+    <bean id="pollableFileSource" class="org.springframework.integration.file.inbound.FileReadingMessageSource"
+          p:directory="#{T(org.springframework.integration.file.inbound.FileReadingMessageSourceIntegrationTests).inputDir}"
           p:filter-ref="compositeFilter"/>
 
 	<bean id="integrationEvaluationContext" class="org.springframework.expression.spel.support.StandardEvaluationContext" />
