+++ spring-integration-file/src/test/java/org/springframework/integration/file/outbound/FileOutboundChannelAdapterIntegrationTests-context.xml
