+++ spring-integration-file/src/test/java/org/springframework/integration/file/outbound/FileOutboundGatewayIntegrationTests-context.xml
